requirements.txt
# ============================================================================
# PYTHON REQUIREMENTS - OPTIMIZED FOR APPLE SILICON (M1/M2/M3)
# ============================================================================
# This file contains dependencies optimized for Apple Silicon with MPS
# (Metal Performance Shaders) GPU acceleration support
# ============================================================================

# Core dependencies
requests>=2.31.0
Pillow>=10.0.0
numpy>=2.0.0

# PyTorch with MPS (Metal Performance Shaders) support
# These versions include native Apple Silicon GPU acceleration
# Install command: pip3 install torch torchvision torchaudio
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers with MPS support
transformers>=4.35.0
accelerate>=0.24.0

# CLIP model dependencies
ftfy>=6.1.0
regex>=2023.0.0
tqdm>=4.65.0

# FastAPI for CLIP service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ============================================================================
# INSTALLATION INSTRUCTIONS:
# ============================================================================
# 1. Create virtual environment:
#    python3 -m venv venv
#
# 2. Activate virtual environment:
#    source venv/bin/activate
#
# 3. Install requirements:
#    pip3 install -r requirements.txt
#
# 4. Verify GPU support:
#    python3 verify_mps.py
#
# Expected: "Apple Silicon GPU acceleration is ENABLED!"
# ============================================================================
